Typescript updates

+++ src/services/moderationReport.ts
@@ -7,7 +7,7 @@ async function getModerationReports({
   status = "open",
 }) {
   const response = await apiClient.get(
-    `/moderation-reports/?limit=${limit}&offset=${offset}&status=${status}`
+    `/moderation-reports/?limit=${limit}&offset=${offset}&status=${status}`,
   );
   return response;
 }
@@ -22,19 +22,20 @@ async function getModerationReportsByUserId({
   userId: string;
 }) {
   const response = await apiClient.get(
-    `/moderation-reports/?limit=${limit}&offset=${offset}&userId=${userId}`
+    `/moderation-reports/?limit=${limit}&offset=${offset}&userId=${userId}`,
   );
   return response;
 }
 
-async function getModerationReportsOnUserId({ userId }: { userId: string }) {
+async function getModerationReportsOnUserId({ userId }: { userId?: string }) {
   const response = await apiClient.get(
-    `/moderation-reports/?reportedUser=${userId}`
+    `/moderation-reports/?reportedUser=${userId}`,
   );
   return response;
 }
 
-async function getModerationReport(id: string) {
+async function getModerationReport(id?: string) {
+  if (!id) return undefined;
   const response = await apiClient.get(`/moderation-reports/${id}`);
   return response;
 }
@@ -42,20 +43,25 @@ async function getModerationReport(id: string) {
 async function postModerationReport(moderationReport: ModerationReport) {
   const response = await apiClient.post(
     `/moderation-reports/`,
-    moderationReport
+    moderationReport,
   );
   return response;
 }
 
-async function putModerationReport(id: string, moderationReport: ModerationReport) {
+async function putModerationReport(
+  id: string | undefined,
+  moderationReport: Partial<ModerationReport>,
+) {
+  if (!id) return undefined;
   const response = await apiClient.put(
     `/moderation-reports/${id}`,
-    moderationReport
+    moderationReport,
   );
   return response;
 }
 
-async function deleteModerationReport(id: string) {
+async function deleteModerationReport(id: string | undefined) {
+  if (!id) return undefined;
   const response = await apiClient.delete(`/moderation-reports/${id}`);
   return response;
 }
